backend/requirements.txt
# Chiron APM v4.0 - Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
orjson>=3.9.0

# Data
snowflake-connector-python>=3.6.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0
httpx>=0.26.0
pytz>=2024.1

# Caching
redis>=5.0.0
cachetools>=5.3.0

# PostgreSQL (local mirror mode)
psycopg2-binary>=2.9.0

# Auth
PyJWT>=2.8.0
